// src/protocols/profile/profile.d.ts
export declare const profile: {
    uri: string;
    schemas: {};
    definition: {
        published: boolean;
        protocol: string;
        types: {
            name: {
                dataFormats: string[];
            };
            social: {
                dataFormats: string[];
            };
            messaging: {
                dataFormats: string[];
            };
            phone: {
                dataFormats: string[];
            };
            address: {
                dataFormats: string[];
            };
            avatar: {
                dataFormats: string[];
            };
            hero: {
                dataFormats: string[];
            };
        };
        structure: {
            social: {};
            avatar: {};
            hero: {};
            name: {};
            messaging: {};
            address: {};
            phone: {};
        };
    };
};